Do not check for overlapping input audio chunks if the Grain is not continuous
Bungee expected overlapping input frame ranges to be equal, even in case of a
requested reset. This doesn't hold when swapping out the input under the hood.
In upstream bungee this would result in logging "UNEXPECTED INPUT" when
instrumentation is on. In our patched version this led to an assertion failure.